third_party/blink/web_tests/external/wpt/webidl/ecmascript-binding/iterator-prototype-object-expected.txt

This is a testharness.js-based test.
PASS Has %IteratorPrototype% as prototype
PASS next() exists and is writable, enumerable, and configurable
FAIL next() throws TypeError when called on ineligible receiver assert_throws_js: function "() => {
    iteratorProto.next.call(new Headers().entries());
  }" did not throw
FAIL Is specific to an interface assert_not_equals: got disallowed value object "[object Iterator]"
Harness: the test ran to completion.
